Changes build mechanism to Grunt and deletes dross
This is a mass mggr-style cull of broken and/or unused code. Primarily
it changes the build mechanism to Grunt, moving away from the python
hand build script to a more commonly used method.

The 'src' folder has been re-jigged and the css and main index.html file
live in here now. To build the application now use `grunt` or for
development use `grunt dev`

Deleted the entire libs folder as it contained things that haven't been
touched in 3 years and were apparently there for the python build script
- very satifying indeed