Added the messages to a template
Made sure that the avitars were updated if they have to be changed to the gravitar icons.

made the avatars a bit bigger and added them to the people list.

Made sure that messages of your own appear on the right hand side.

removed the 3D effects

Made each message from the same person join to their previous mesage if they are not interupted